+++ src/main/java/net/fabricmc/mappingio/adapter/OuterClassNamePropagator.java
@@ -17,7 +17,9 @@
 package net.fabricmc.mappingio.adapter;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -34,8 +36,9 @@ import net.fabricmc.mappingio.MappingUtil;
 import net.fabricmc.mappingio.MappingVisitor;
 
 /**
- * Searches for inner classes with no mapped name, whose enclosing classes do have mapped names,
- * and applies those to the outer part of the inner classes' fully qualified name.
+ * Searches for inner classes whose effective destination name contains outer classes referenced via their source name,
+ * waits for mappings for these enclosing classes, and applies the latter's destination names
+ * to the former's fully qualified name.
  *
  * <p>For example, it takes a class {@code class_1$class_2} that doesn't have a mapping,
  * tries to find {@code class_1}, which let's say has the mapping {@code SomeClass},
@@ -45,8 +48,22 @@ import net.fabricmc.mappingio.MappingVisitor;
  * the other to actually apply the outer names. The third pass onwards will then emit the final mappings.
  */
 public class OuterClassNamePropagator extends ForwardingMappingVisitor {
+	/**
+	 * Constructs a new {@link OuterClassNamePropagator} which processes all destination namespaces,
+	 * including already remapped class destination names therein.
+	 */
 	public OuterClassNamePropagator(MappingVisitor next) {
+		this(next, null, true);
+	}
+
+	/**
+	 * @param namespaces The destination namespaces where outer class names shall be propagated (isolated from each other).
+	 * @param processRemappedDstNames Whether already remapped destination names should also get their unmapped outer classes replaced.
+	 */
+	public OuterClassNamePropagator(MappingVisitor next, Collection<String> namespaces, boolean processRemappedDstNames) {
 		super(next);
+		this.dstNamespacesToProcess = namespaces;
+		this.processRemappedDstNames = processRemappedDstNames;
 	}
 
 	@Override
@@ -60,7 +77,7 @@ public class OuterClassNamePropagator extends ForwardingMappingVisitor {
 
 	@Override
 	public boolean visitHeader() throws IOException {
-		if (pass < firstEmitPass) return true;
+		if (pass < FIST_EMIT_PASS) return true;
 
 		return super.visitHeader();
 	}
@@ -68,26 +85,53 @@ public class OuterClassNamePropagator extends ForwardingMappingVisitor {
 	@Override
 	@SuppressWarnings("unchecked")
 	public void visitNamespaces(String srcNamespace, List<String> dstNamespaces) throws IOException {
-		dstNsCount = dstNamespaces.size();
+		if (pass == COLLECT_CLASSES_PASS) {
+			if (dstNamespacesToProcess == null) {
+				dstNamespacesToProcess = dstNamespaces;
+			} else {
+				if (dstNamespacesToProcess.contains(srcNamespace)) {
+					throw new UnsupportedOperationException(srcNamespace + " was passed as a destination namespace"
+							+ " to propagate outer class names to, but has been visited as the source namespace.");
+				}
 
-		if (pass == collectClassesPass) {
+				for (String ns : dstNamespacesToProcess) {
+					if (!dstNamespaces.contains(ns)) {
+						throw new IllegalArgumentException(ns + " was passed as a destination namespace to propagate outer class names to,"
+								+ " but is not present in the namespaces of the current visitation pass.");
+					}
+				}
+			}
+
+			this.dstNamespaces = dstNamespaces;
+			this.dstNsCount = dstNamespaces.size();
+
+			if (dstNamespaceIndicesToProcess == null) {
+				dstNamespaceIndicesToProcess = new ArrayList<>();
+
+				for (int i = 0; i < dstNsCount; i++) {
+					if (dstNamespacesToProcess.contains(dstNamespaces.get(i))) {
+						dstNamespaceIndicesToProcess.add(i);
+					}
+				}
+			}
+
 			visitedDstName = new boolean[dstNsCount];
 			dstNameBySrcNameByNamespace = new HashMap[dstNsCount];
-		} else if (pass >= firstEmitPass) {
+		} else if (pass >= FIST_EMIT_PASS) {
 			super.visitNamespaces(srcNamespace, dstNamespaces);
 		}
 	}
 
 	@Override
 	public void visitMetadata(String key, @Nullable String value) throws IOException {
-		if (pass < firstEmitPass) return;
+		if (pass < FIST_EMIT_PASS) return;
 
 		super.visitMetadata(key, value);
 	}
 
 	@Override
 	public boolean visitContent() throws IOException {
-		if (pass < firstEmitPass) return true;
+		if (pass < FIST_EMIT_PASS) return true;
 
 		return super.visitContent();
 	}
@@ -96,9 +140,9 @@ public class OuterClassNamePropagator extends ForwardingMappingVisitor {
 	public boolean visitClass(String srcName) throws IOException {
 		this.srcName = srcName;
 
-		if (pass == collectClassesPass) {
+		if (pass == COLLECT_CLASSES_PASS) {
 			dstNamesBySrcName.putIfAbsent(srcName, new String[dstNsCount]);
-		} else if (pass >= firstEmitPass) {
+		} else if (pass >= FIST_EMIT_PASS) {
 			super.visitClass(srcName);
 		}
 
@@ -107,11 +151,11 @@ public class OuterClassNamePropagator extends ForwardingMappingVisitor {
 
 	@Override
 	public void visitDstName(MappedElementKind targetKind, int namespace, String name) throws IOException {
-		if (pass == collectClassesPass) {
+		if (pass == COLLECT_CLASSES_PASS) {
 			if (targetKind != MappedElementKind.CLASS) return;
 
 			dstNamesBySrcName.get(srcName)[namespace] = name;
-		} else if (pass >= firstEmitPass) {
+		} else if (pass >= FIST_EMIT_PASS) {
 			if (targetKind == MappedElementKind.CLASS) {
 				visitedDstName[namespace] = true;
 				name = dstNamesBySrcName.get(srcName)[namespace];
@@ -123,7 +167,7 @@ public class OuterClassNamePropagator extends ForwardingMappingVisitor {
 
 	@Override
 	public void visitDstDesc(MappedElementKind targetKind, int namespace, String desc) throws IOException {
-		if (pass < firstEmitPass) return;
+		if (pass < FIST_EMIT_PASS) return;
 
 		if (modifiedClasses.contains(srcName)) {
 			Map<String, String> nsDstNameBySrcName = dstNameBySrcNameByNamespace[namespace];
@@ -143,23 +187,37 @@ public class OuterClassNamePropagator extends ForwardingMappingVisitor {
 
 	@Override
 	public boolean visitElementContent(MappedElementKind targetKind) throws IOException {
-		if (targetKind == MappedElementKind.CLASS && pass > collectClassesPass) {
+		if (targetKind == MappedElementKind.CLASS && pass > COLLECT_CLASSES_PASS) {
 			String[] dstNames = dstNamesBySrcName.get(srcName);
 
 			for (int ns = 0; ns < dstNames.length; ns++) {
+				if (!dstNamespacesToProcess.contains(dstNamespaces.get(ns))) {
+					continue;
+				}
+
 				String dstName = dstNames[ns];
 
-				if (pass == fixOuterClassesPass) {
-					if (dstName != null) continue; // skip if already mapped
+				if (pass == FIX_OUTER_CLASSES_PASS) {
+					if (!processRemappedDstNames && dstName != null && !dstName.equals(srcName)) {
+						continue;
+					}
 
-					String[] parts = srcName.split(Pattern.quote("$"));
+					String[] srcParts = srcName.split(Pattern.quote("$"));
+					String[] dstParts = dstName == null ? srcParts : dstName.split(Pattern.quote("$"));
+					assert dstParts.length == srcParts.length;
 
-					for (int pos = parts.length - 2; pos >= 0; pos--) {
-						String outerSrcName = String.join("$", Arrays.copyOfRange(parts, 0, pos + 1));
+					for (int pos = srcParts.length - 2; pos >= 0; pos--) {
+						String outerSrcName = String.join("$", Arrays.copyOfRange(srcParts, 0, pos + 1));
+
+						if (dstName != null && !dstParts[pos].equals(srcParts[pos])) {
+							// That part already has a differing mapping
+							continue;
+						}
+
 						String outerDstName = dstNamesBySrcName.get(outerSrcName)[ns];
 
-						if (outerDstName != null) {
-							dstName = outerDstName + "$" + String.join("$", Arrays.copyOfRange(parts, pos + 1, parts.length));
+						if (outerDstName != null && !outerDstName.equals(outerSrcName)) {
+							dstName = outerDstName + "$" + String.join("$", Arrays.copyOfRange(dstParts, pos + 1, dstParts.length));
 
 							dstNames[ns] = dstName;
 							modifiedClasses.add(srcName);
@@ -176,7 +234,7 @@ public class OuterClassNamePropagator extends ForwardingMappingVisitor {
 			}
 		}
 
-		if (pass < firstEmitPass) {
+		if (pass < FIST_EMIT_PASS) {
 			return false; // prevent other element visits, we only care about classes here
 		}
 
@@ -186,18 +244,22 @@ public class OuterClassNamePropagator extends ForwardingMappingVisitor {
 
 	@Override
 	public boolean visitEnd() throws IOException {
-		if (pass++ < firstEmitPass) {
+		if (pass++ < FIST_EMIT_PASS) {
 			return false;
 		}
 
 		return super.visitEnd();
 	}
 
-	private static final int collectClassesPass = 1;
-	private static final int fixOuterClassesPass = 2;
-	private static final int firstEmitPass = 3;
+	private static final int COLLECT_CLASSES_PASS = 1;
+	private static final int FIX_OUTER_CLASSES_PASS = 2;
+	private static final int FIST_EMIT_PASS = 3;
+	private final boolean processRemappedDstNames;
 	private final Map<String, String[]> dstNamesBySrcName = new HashMap<>();
 	private final Set<String> modifiedClasses = new HashSet<>();
+	private List<String> dstNamespaces;
+	private Collection<String> dstNamespacesToProcess;
+	private Collection<Integer> dstNamespaceIndicesToProcess;
 	private int pass = 1;
 	private int dstNsCount = -1;
 	private String srcName;